Object and ELF files seems to now contain valid-ish DWARF data. Two byte addresses are now acceptable for DWARF code.
Updated DWARF test suite correspondingly.

Register names are now prefixed with llvm_mos_ and are now printed
that way, to avoid conflicts with existing code that depends on a, x, y, sp
and the like.

A feature, mos-long-register-names, has been added that requires that kind of
naming convention; by default, the long register names are not required.