Unify AArch64 tests into a single file (1/4) (NFC) (#146328)
This is a series of patches (1/4) to unify assembly/disassembly of
recent AArch64 tests into a single file. The aim is to improve
consistency, so that all instructions and system registers are
thoroughly tested, and future test cases will be in a unified format.

This patch:
 * unifies errorless .s and .txt tests into a single file
 * remove .txt tests which don't have feature requirements
* makes the .s tests have a roundabout run line to test both encoding
and assembly

See also #146329, #146330 and #146331.

---------

Co-authored-by: Virginia Cangelosi <virginia.cangelosi@arm.com>