Replace SIMDLongThreeVectorBHSabd with SIMDLongThreeVectorBHS. (#152987)
We just need to use a BinOpFrag to share the patterns. This also moves
UABDL to where it belongs in with similar instructions, and removes some
patterns that are now handled by abd nodes. This is mostly NFC except
for GISel, which will catch back up when it handles abd nodes in the
same way.