Reporting Errors for Duplicating Link Declarations in `modulemap`s (#148959)
This PR teaches the modulemap parsing logic to report warnings that
default to errors if the parsing logic sees duplicating link
declarations in the same module. Specifically, duplicating link
declarations means multiple link declarations with the same
string-literal in the same module. No errors are reported if a same link
declaration exist in a submodule and its enclosing module.

The warning can be disabled with `-Wno-module-link-redeclaration`.

rdar://155880064