Skip system macros in readability-identifier-naming check (#132016)
Currently, the check is processing system macros. Most importantly, it
tries to find .clang-tidy files associated with those files, if the
option GetConfigPerFile (on by default) is active.

This is problematic for a number of reasons:

- System macros cannot be acted upon (renamed), so it's wasted work.
- When the main .cpp file includes a system header, clang-tidy tries to
find the .clang-tidy file for that system header. When that system
header is a 3rd-party repository, they may have their own .clang-tidy
file, which may not be compatible with the current version of
clang-tidy. So, clang-tidy may fail to analyze our main.cpp file, only
because it includes a 3rd-party system header whose .clang-tidy file is
incompatible with our clang-tidy binary.

Therefore, skip system macros in this check.

This reduces the amount of unwanted warnings as follows:

On trunk:
Suppressed 11448 warnings (11448 in non-user code).

With this patch:
Suppressed 4824 warnings (4824 in non-user code).

The runtime is pretty much unchanged.

Co-authored-by: Carlos Gálvez <carlos.galvez@zenseact.com>