Remove leftovers of FreeBSD md5/sha2 interceptors (#153351)
This code was added in 18a7ebda99044473fdbce6376993714ff54e6690, but was
not properly reverted in d0b9c2c5647656738cda3fb670aa5d3b3a69d784.

Removing this code makes it possible to cross-compile TSan for FreeBSD
with `zig cc` which does not provide libcrypt when cross-compiling.

I also removed a `term.h` include (ncurses) which appeared to be
unnecessary and likewise prevented cross-compilation with `zig cc` from
working.