Add C tests for recipe generation, fix NYI
I realized while messing with other things that I'd written all of the
recipe tests for C++, so this patch adds a bunch of tests for C mode.
The assert wasn't quite accurate (as C default init doesn't really do
anything/have an AST node), so that is corrected.  Also, the lack of
cir.copy causes some of the firstprivate tests to be incomplete, so
added TODOs for that as well.