Produce canonical linalg.generic for im2col (#134675)
Before this patch, the Img2Col transform produced a non-canonical
linalg.generic whose input tensor was not reported in the inputs of the
operation: instead, it was accessed manually from inside the op body,
after an internal calculation of the access offsets. This patch modifies
the Im2Col rewrite to produce a canonical linalg.generic whose input is
correctly reported in its 'ins()', whose access offsets are computed
through an indexing map, and whose body contains only a 'linalg.yield'
op.

Signed-off-by: Fabrizio Indirli <Fabrizio.Indirli@arm.com>
Co-authored-by: Georgios Pinitas <georgios.pinitas@arm.com>