Added storage specification for [hl]fir.declare. (#155325)
As proposed in
https://discourse.llvm.org/t/rfc-flang-representation-for-objects-inside-physical-storage/88026,
this patch adds a `storage` Value operand and a `storage_offset`
Integer attribute for `[hl]fir.declare` operations.

The `storage` operand indicates the raw address of the physical storage
a variable belongs to. This is the beginning address of the physical
storage.
The `storage_offset` specifies a byte offset within the physical storage
where the variable object starts.