fix nested tags of the same name not being included in their context (#155965)
Fix an error in the logic meant to handle a redeclaration such as:
```C++
struct A {
  struct __attribute__((foo)) A *ptr;
};
```
In the declaration of ptr, we must introduce a new redeclaration of A in
order for it to carry the new attribute. This is a redeclaration of the
existing A, but it is only lexically contained in A, still semantically
belonging to the TU. This is the same deal as happens with friend
declarations, and the logic used to handle that is reused here.

But this was going haywire with a class indirectly nested within a class
of the same name.

The fix limits this logic to only apply when the tag use is just a
simple reference.

Since this regression was never released, there are no release notes.

Fixes #155936