Fix of a crash 'Cannot get layout of forward declarations' during CTU static analysis (#155375)
When a type is imported with `ASTImporter`, the "original declaration"
of the type is imported. In some cases this is not the definition (of
the class). Before the fix the definition was only imported if there was
an other reference to it in the AST to import. This is not always the
case (like in the added test case), if not the definition was missing in
the "To" AST which can cause the assertion later.