Fix tests that were unresolved when using lit's internal shell
These two tests were unresolved when using lit's internal shell.

In the case of tail-duplication-constant-prop, it was because they were
using a echo $? line, and lit's internal echo implementation does not
support $? to get the return code. The test was never actually asserting
anything about the return code though, so I've removed the echo
commands.

In the case of permission.test, it was because umask was not supported
before #155850, and afterwards not without an argument. The test also
was not great at capturing what it was supposed to (leaving open
possibilites like the system umask and what Bolt was using happening to
match), so I've rewritten the test in the style of
llvm/test/tools/llvm-objcopy/ELF/respect-umask.test.

This fixes #102693.

Reviewers:
maksfb, yota9, ayermolo, yozhu, aaupov, rafaelauler, paschalis-mpeis

Reviewed By: maksfb

Pull Request: https://github.com/llvm/llvm-project/pull/156082