Use target global AS for SrcLocStr (#156520)
We should set the correct target-specific AS for the SrcLocStr global
created in OMPIRBuilder.

We also may have to insert a constexpr addrspacecast because the struct
field type may be different than the value used to initialize it.

I actually want the cast to be from AS 1 to AS 4, but getting the type
to be AS4 relies on a PR currently in-review, so leave the cast target
to AS 0 for now.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>