Add support for delegating constructor initialization (#156757)
This adds support for zero-initialization during delegating constructor
processing.

Note, this also adds code to skip emitting constructors that are trivial
and default to match the classic codegen behavior. The incubator does
not skip these constructors, but I have found a case where this results
in a call to a default constructor that is never defined.