canonicalize shape in pointer remapping lowering (#157472)
Fix for #153221.

Canonicalize the new shape of the pointer when lowering pointer
assignment with bounds remapping.
This is done by using the existing helper that generates a compare to
zero + select like in the other situation where shapes are lowered.

Note that this only needs to be done for the extents here because lower
bounds are canonicalized in LBOUND inquiries and descriptor creation
(embox/rebox codegen) based on the extent value.