Rewrite stdin.s to use python
This test needs to change the file descriptor offset for the llvm-mc
output file to test that we do not get an assertion in that situation.
This doesn't seem easy to do without bash subshells. Rewrite these test
in Python so we can remove the shell requirement from this test and
enable using it with lit's internal shell. This also has the bonus of
making the behavior that we are trying to create for the test much more
explicit (a .seek call on the FD).

Reviewers: ilovepi, MaskRay, petrhosek

Reviewed By: petrhosek, ilovepi

Pull Request: https://github.com/llvm/llvm-project/pull/157232