Change sig-trap test to allow expected output strings in any order (#157407)
I have seen some flakiness in this test where the 2 checked strings
appear in a different order. Due to buffering of writes, and that one of
these strings is written during the signal handler, I think this is
valid.

This PR relaxes the test to allow those strings to appear in
either order.