Fix __cpuidex conflict with other offloading targets (#157741)
It seems that for whatever reason we must:
1) Declare aux builtins when the compiling for an offload device
and
2) Define the aux builtin target macros when compiling for an offload
device.

In `cpuid.h` we try to define `__cpuidex` if it is not defined. Given
the above, the function will both be defined as a builtin in the
compiler and we can't rely on the `X86` macros to be undefined in the
case the aux-triple is `X86`.

Previously a workaround was added for NVPTX in
https://github.com/llvm/llvm-project/pull/152556, extend it for the
other offloading targets.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>