Remove commented install line and disutils reference (#158015)
76667c768e6403e71718340a946e6f2f356bf745 added distutils because "spawn"
was used, which I then removed in
268a4b0a451432833d3b398c7182d133c865dff5.

I removed it as part of removing all uses of distutils in llvm-project,
tracked in #54337. Python has removed distutils in its latest versions.

Distutils was not being installed in the docker image but just mentioned
in a commented out line. I think this line was leftover from when it was
reformated into the multi-line command above.

So I'm removing the whole line and relocating the comments a bit.