Loosen Clang driver test regexes to fix buildbots (#159158)
Make the test regexes more permissive to fix buildbot failures caused by
the merge of PR #159129. This mirrors the earlier fix in PR #148908. We
retain cross-project-test coverage to verify that the path content is
correct.

A follow-up will update the tests to robustly check the Clang executable
filename, likely along the lines of PR #159151.

Short-term unbreak to keep the buildbots green without reverting a chain
of commits.