Update IssueWidth for Neoverse V1, N1, N3 (#154495)
Recently the IssueWidth in the scheduling model was reduced for
Neoverse-V2 and N2. This patch does the same for Neoverse-V1, N1 and N3.

On Neoverse-V1, various values of IssueWidth (15, 8, 7, 6, 5) were tried
with runs of various workloads.
The highest overall geomean score was achieved with an issue width of 8.
No significant regressions were noted.

On Neoverse-N1, various values of IssueWidth (8, 6, 5, 4, 3) were tried
with runs of various workloads.
The highest overall geomean score was achieved with an issue width of 3.
No significant regressions were noted.

On Neoverse-N3, it makes sense to do exactly the same as was done for
N2. It is proposed to use an issue width of 5.

Related V2 PR: https://github.com/llvm/llvm-project/pull/142565
Related N2 PR: https://github.com/llvm/llvm-project/pull/145717