Add support for readfile to external shell
This patch adds support for the new lit %{readfile:<filename>}
substitution to the external shell. The implementation currently just
appends some test commands to ensure the file exists and uses a subshell
with cat. This is intended to enable running tests using the
substitution in the external shell before we fully switch over to the
internal shell.

This code is designed to be temporary with us deleting it once
everything has migrated over to the internal shell and we are able to
remove the external shell code paths.

Reviewers: petrhosek, cmtice, pogo59, ilovepi, arichardson

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-project/pull/159431