Make builtin cat work with stdin
cat with no files passed to it is supposed to read from STDIN according
to POSIX. The builtin cat lacking this behavior led to the clang test in
dev-fd-fs.c to fail because it expected this behavior. This is a simple
modification and I do not think it is possible to rewrite the test
without this easily while preserving the semantics around named pipes.

Reviewers: petrhosek, arichardson, ilovepi, cmtice, jh7370

Reviewed By: jh7370, arichardson, ilovepi, cmtice

Pull Request: https://github.com/llvm/llvm-project/pull/158447