Robustify Tosa_IfOp against null dereference and wrong assertion (#159756)
Fixes  #159650.

The current implementation ICE out if we access an IfOp's terminator
when it doesn't have it. Instead the PR defers the job of verifying that
a block would have at least a terminator.

The current implementation also crashes with cast<YieldOp> if the
terminator is not a YieldOp, the PR also defers the job of verification
to the op itself.