Remove bfloat vectors from AllFloatVectors. NFC (#160399)
This removes a bunch of unreachable isel patterns for floating point
operations like fadd, fsub, fmul, etc.

Eventually we will need patterns for Zvfbfa but this will require new
pseudoinstructions with the altfmt bit set so these extra patterns
aren't helpful for that either.

Add a new AllFloatAndBFloatVectors for the instructions that we do need
both for like vrgather, vcompress, vmerge.