Don't use ResultElem for assumption list (NFC) (#160462)
ResultElem stores a weak handle of an assume, plus an index for
referring to a specific operand bundle. This makes sense for the results
of assumptionsFor(), which refers to specific operands of assumes.

However, assumptions() is a plain list of assumes. It does *not* contain
separate entries for each operand bundles. The operand bundle index is
always ExprResultIdx.

As such, we should be directly using WeakVH for this case, without the
additional wrapper.