Refactor MCInstReference and move it to Core (NFC) (#155846)
Refactor MCInstReference class and move it from PAuth gadget scanner to
Core.

MCInstReference is a class representing a reference to a constant
instruction inside a parent entity - either inside a basic block (which
has a reference to its parent function) or directly inside a function
(when CFG information is not available).

This patch reapplies #138655 with a fix for iterator usage and multiple
minor issues fixed during the second round of review.